docs: Approach2
Using two mutable state objects, "stage" and "lemonSqueeze".

Stage - tracks which part of the process we are in.
lemonSqueeze - counts the number of times lemon is squeezed.

Using these two, we will assign values to the UI elements in  "LemonadeUI", a composable function that defines all our UI.
I have yet to define all UI elements.